add logic for changing current health in the in-game UI

+++ Assets/Scripts/SO/PlayerSettingsSO.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "PlayerSettings", menuName = "ScriptableObjects/PlayerSettings")]
 public class PlayerSettingsSO : ScriptableObject
 {
-    public float Speed;
+    public int Speed;
     public float RotationSpeed;
     public int Health;
     public int Damage ;
